[Upd] fixes

+++ Server/includes/listener.hpp
@@ -10,7 +10,7 @@ class Listener{
         uint8_t reception_type();
         int reception();
         int envoie_bool(uint8_t type_msg ,uint8_t boolint);
-        int envoie_msg(int type_msg , std::string msg);
+        int envoie_msg(uint8_t type_msg , std::string msg);
         char* get_buffer();
 
     private:
